Version 1.6: Major overhaul of #reaper-crypto!
What's New
In addition to just bitcoin, the cryptocurrency ethereum is now supported as well! The channel changed from "reaper-bitcoin" to "reaper-crypto" to reflect this.
The syntax on the commands has changed slightly, check it out!
The bot handles commands given with incorrect parameters on various functions (begin game,buy, etc.) better now.
The price of each cryptocurrency updates much faster now, approximately every minute!
Major overhaul to the code, behind the scenes.

Bug Fixes
Stopped the simulator from crashing when NaN is passed in as a parameter.